Add jest-fetch-mock plugin, move getRandomIdFromArray function to util.js file
